Handled some warnings and optimized imports

+++ cyfra-foton/src/main/scala/io/computenode/cyfra/foton/animation/AnimationRenderer.scala
@@ -2,21 +2,14 @@ package io.computenode.cyfra.foton.animation
 
 import io.computenode.cyfra
 import io.computenode.cyfra.dsl.Value.*
-import io.computenode.cyfra.foton.rt.ImageRtRenderer.RaytracingIteration
-import io.computenode.cyfra.foton.rt.animation.AnimationRtRenderer.RaytracingIteration
-import io.computenode.cyfra.foton.rt.RtRenderer
-import io.computenode.cyfra.foton.rt.animation.AnimatedScene
+import io.computenode.cyfra.dsl.{*, given}
 import io.computenode.cyfra.runtime.GFunction
+import io.computenode.cyfra.runtime.mem.GMem.fRGBA
+import io.computenode.cyfra.utility.ImageUtility
 import io.computenode.cyfra.utility.Units.Milliseconds
 import io.computenode.cyfra.utility.Utility.timed
-import io.computenode.cyfra.{*, given}
-import io.computenode.cyfra.utility.ImageUtility
-import io.computenode.cyfra.dsl.{*, given}
-import io.computenode.cyfra.runtime.mem.GMem.fRGBA
 
-import java.nio.file.{Path, Paths}
-import scala.concurrent.Await
-import scala.concurrent.duration.DurationInt
+import java.nio.file.Path
 
 trait AnimationRenderer[S <: AnimationRenderer.Scene, F <: GFunction[?, Vec4[Float32], Vec4[Float32]]](params: AnimationRenderer.Parameters):
 
@@ -25,7 +18,7 @@ trait AnimationRenderer[S <: AnimationRenderer.Scene, F <: GFunction[?, Vec4[Flo
   def renderFramesToDir(scene: S, destinationPath: Path): Unit =
     destinationPath.toFile.mkdirs()
     val images = renderFrames(scene)
-    val totalFrames = Math.ceil(scene.duration.toFloat / msPerFrame).toInt
+    val totalFrames = Math.ceil(scene.duration / msPerFrame).toInt
     val requiredDigits = Math.ceil(Math.log10(totalFrames)).toInt
     images.zipWithIndex.foreach:
       case (image, i) =>
@@ -35,7 +28,7 @@ trait AnimationRenderer[S <: AnimationRenderer.Scene, F <: GFunction[?, Vec4[Flo
 
   def renderFrames(scene: S): LazyList[Array[fRGBA]] =
     val function = renderFunction(scene)
-    val totalFrames = Math.ceil(scene.duration.toFloat / msPerFrame).toInt
+    val totalFrames = Math.ceil(scene.duration / msPerFrame).toInt
     val timestamps = LazyList.range(0, totalFrames).map(_ * msPerFrame)
     timestamps.zipWithIndex.map { case (time, frame) =>
       timed(s"Animated frame $frame/$totalFrames"):
